Use protokit utils in test
