+++ Gopkg.toml
@@ -35,4 +35,4 @@
 
 [[constraint]]
   name = "github.com/pseudomuto/protokit"
-  version = "0.1.0-pre"
+  version = "0.1.0-pre2"
