+++ fixtures/generate.go
@@ -1,5 +1,3 @@
 package main
 
-//go:generate go build -o ./fixtures main.go
-//go:generate protoc --plugin=protoc-gen-fix=./fixtures -I=./protos --fix_out=. -I../ -I../options ./protos/doc.proto ./protos/todo/service.proto ./protos/greeter/service.proto
-//go:generate rm fixtures
+//go:generate protoc --descriptor_set_out=fileset.pb --include_imports --include_source_info -I. -I../ -I../options doc.proto todo/service.proto greeter/service.proto
